+++ frontend/app/.server/constants/types.constant.ts
@@ -38,14 +38,13 @@ import type {
   ClientFriendlyStatusRepository,
   CountryRepository,
   DemographicSurveyRepository,
-  FederalGovernmentInsurancePlanRepository,
+  FederalProvincialGovernmentInsurancePlanRepository,
   LetterRepository,
   LetterTypeRepository,
   MaritalStatusRepository,
   PreferredCommunicationMethodRepository,
   PreferredLanguageRepository,
   ProvinceTerritoryStateRepository,
-  ProvincialGovernmentInsurancePlanRepository,
   VerificationCodeRepository,
 } from '~/.server/domain/repositories';
 import type {
@@ -177,14 +176,13 @@ export const TYPES = assignServiceIdentifiers({
       ClientFriendlyStatusRepository: serviceId<ClientFriendlyStatusRepository>(),
       CountryRepository: serviceId<CountryRepository>(),
       DemographicSurveyRepository: serviceId<DemographicSurveyRepository>(),
-      FederalGovernmentInsurancePlanRepository: serviceId<FederalGovernmentInsurancePlanRepository>(),
+      FederalProvincialGovernmentInsurancePlanRepository: serviceId<FederalProvincialGovernmentInsurancePlanRepository>(),
       LetterRepository: serviceId<LetterRepository>(),
       LetterTypeRepository: serviceId<LetterTypeRepository>(),
       MaritalStatusRepository: serviceId<MaritalStatusRepository>(),
       PreferredCommunicationMethodRepository: serviceId<PreferredCommunicationMethodRepository>(),
       PreferredLanguageRepository: serviceId<PreferredLanguageRepository>(),
       ProvinceTerritoryStateRepository: serviceId<ProvinceTerritoryStateRepository>(),
-      ProvincialGovernmentInsurancePlanRepository: serviceId<ProvincialGovernmentInsurancePlanRepository>(),
       VerificationCodeRepository: serviceId<VerificationCodeRepository>(),
     },
     services: {
