+++ frontend/app/.server/domain/repositories/federal-provincial-government-insurance-plan.repository.ts
@@ -1,0 +1,127 @@
+import { injectable } from 'inversify';
+
+import type { FederalGovernmentInsurancePlanEntity, ProvincialGovernmentInsurancePlanEntity } from '~/.server/domain/entities';
+import { createLogger } from '~/.server/logging';
+import type { Logger } from '~/.server/logging';
+import federalGovernmentInsurancePlanJsonDataSource from '~/.server/resources/power-platform/federal-government-insurance-plan.json';
+import provincialGovernmentInsurancePlanJsonDataSource from '~/.server/resources/power-platform/provincial-government-insurance-plan.json';
+
+export interface FederalProvincialGovernmentInsurancePlanRepository {
+  /**
+   * Fetch all federal government insurance plan entities.
+   * @returns All federal government insurance plan entities.
+   */
+  listAllFederalGovernmentInsurancePlans(): FederalGovernmentInsurancePlanEntity[];
+
+  /**
+   * Fetch a federal government insurance plan entity by its id.
+   * @param id The id of the federal government insurance plan entity.
+   * @returns The federal government insurance plan entity or null if not found.
+   */
+  findFederalGovernmentInsurancePlanById(id: string): FederalGovernmentInsurancePlanEntity | null;
+
+  /**
+   * Fetch all provincial government insurance plan entities.
+   * @returns All provincial government insurance plan entities.
+   */
+  listAllProvincialGovernmentInsurancePlans(): ProvincialGovernmentInsurancePlanEntity[];
+
+  /**
+   * Fetch a provincial government insurance plan entity by its id.
+   * @param id The id of the provincial government insurance plan entity.
+   * @returns The provincial government insurance plan entity or null if not found.
+   */
+  findProvincialGovernmentInsurancePlanById(id: string): ProvincialGovernmentInsurancePlanEntity | null;
+}
+
+@injectable()
+export class DefaultFederalProvincialGovernmentInsurancePlanRepository implements FederalProvincialGovernmentInsurancePlanRepository {
+  private readonly log: Logger;
+
+  constructor() {
+    this.log = createLogger('DefaultFederalProvincialGovernmentInsurancePlanRepository');
+  }
+
+  listAllFederalGovernmentInsurancePlans(): FederalGovernmentInsurancePlanEntity[] {
+    throw new Error('Federal government insurance plan service is not yet implemented');
+    //TODO: Implement listAllFederalGovernmentInsurancePlans service
+  }
+
+  findFederalGovernmentInsurancePlanById(id: string): FederalGovernmentInsurancePlanEntity | null {
+    throw new Error('Federal government insurance plan service is not yet implemented');
+    //TODO: Implement findFederalGovernmentInsurancePlanById service
+  }
+
+  listAllProvincialGovernmentInsurancePlans(): ProvincialGovernmentInsurancePlanEntity[] {
+    throw new Error('Provincial government insurance plan service is not yet implemented');
+    //TODO: Implement listAllProvincialGovernmentInsurancePlans service
+  }
+
+  findProvincialGovernmentInsurancePlanById(id: string): ProvincialGovernmentInsurancePlanEntity | null {
+    throw new Error('Provincial government insurance plan service is not yet implemented');
+    //TODO: Implement findProvincialGovernmentInsurancePlanById service
+  }
+}
+
+@injectable()
+export class MockFederalProvincialGovernmentInsurancePlanRepository implements FederalProvincialGovernmentInsurancePlanRepository {
+  private readonly log: Logger;
+
+  constructor() {
+    this.log = createLogger('MockFederalProvincialGovernmentInsurancePlanRepository');
+  }
+
+  listAllFederalGovernmentInsurancePlans(): FederalGovernmentInsurancePlanEntity[] {
+    this.log.debug('Fetching all federal government insurance plans');
+    const federalGovernmentInsurancePlanEntities = federalGovernmentInsurancePlanJsonDataSource.value;
+
+    if (federalGovernmentInsurancePlanEntities.length === 0) {
+      this.log.warn('No federal government insurance plans found');
+      return [];
+    }
+
+    this.log.trace('Returning federal government insurance plans: [%j]', federalGovernmentInsurancePlanEntities);
+    return federalGovernmentInsurancePlanEntities;
+  }
+
+  findFederalGovernmentInsurancePlanById(id: string): FederalGovernmentInsurancePlanEntity | null {
+    this.log.debug('Fetching federal government insurance plan with id: [%s]', id);
+
+    const federalGovernmentInsurancePlanEntities = federalGovernmentInsurancePlanJsonDataSource.value;
+    const federalGovernmentInsurancePlanEntity = federalGovernmentInsurancePlanEntities.find(({ esdc_governmentinsuranceplanid }) => esdc_governmentinsuranceplanid === id);
+
+    if (!federalGovernmentInsurancePlanEntity) {
+      this.log.info('Federal government insurance plan not found; id: [%s]', id);
+      return null;
+    }
+
+    return federalGovernmentInsurancePlanEntity;
+  }
+
+  listAllProvincialGovernmentInsurancePlans(): ProvincialGovernmentInsurancePlanEntity[] {
+    this.log.debug('Fetching all provincial government insurance plans');
+    const provincialGovernmentInsurancePlanEntities = provincialGovernmentInsurancePlanJsonDataSource.value;
+
+    if (provincialGovernmentInsurancePlanEntities.length === 0) {
+      this.log.warn('No provincial government insurance plans found');
+      return [];
+    }
+
+    this.log.trace('Returning provincial government insurance plans: [%j]', provincialGovernmentInsurancePlanEntities);
+    return provincialGovernmentInsurancePlanEntities;
+  }
+
+  findProvincialGovernmentInsurancePlanById(id: string): ProvincialGovernmentInsurancePlanEntity | null {
+    this.log.debug('Fetching provincial government insurance plan with id: [%s]', id);
+
+    const provincialGovernmentInsurancePlanEntities = provincialGovernmentInsurancePlanJsonDataSource.value;
+    const provincialGovernmentInsurancePlanEntity = provincialGovernmentInsurancePlanEntities.find(({ esdc_governmentinsuranceplanid }) => esdc_governmentinsuranceplanid === id);
+
+    if (!provincialGovernmentInsurancePlanEntity) {
+      this.log.info('Provincial government insurance plan not found; id: [%s]', id);
+      return null;
+    }
+
+    return provincialGovernmentInsurancePlanEntity;
+  }
+}
