+++ frontend/app/.server/domain/repositories/index.ts
@@ -8,12 +8,11 @@ export * from './client-application.repository';
 export * from './client-friendly-status.repository';
 export * from './country.repository';
 export * from './demographic-survey.repository';
-export * from './federal-government-insurance-plan.repository';
 export * from './letter-type.repository';
 export * from './letter.repository';
 export * from './marital-status.repository';
 export * from './preferred-communication-method.repository';
 export * from './preferred-language.repository';
 export * from './province-territory-state.repository';
-export * from './provincial-government-insurance-plan.repository';
+export * from './federal-provincial-government-insurance-plan.repository';
 export * from './verification-code.repository';
